% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdr_data.R
\name{get_hdr_data}
\alias{get_hdr_data}
\title{Fetch data from HDR API}
\usage{
get_hdr_data(
  country_or_aggregation = NULL,
  year = NULL,
  indicator = NULL,
  apikey,
  query = NULL
)
}
\arguments{
\item{country_or_aggregation}{}

\item{year}{numeric}

\item{indicator}{str}

\item{apikey}{str}

\item{query}{default NULL, else 'detailed' for complete information}
}
\value{
a dataframe
}
\description{
Fetch data from HDR API
}
\examples{
get_hdr_data(country_or_aggregation = 'all_countries', apikey = HDR_API, indicator = 'co2_prod', query = 'detailed')
}
