% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get metadata from the human development report API}
\usage{
get_meta(value, apikey)
}
\arguments{
\item{value}{one from 'Countries', 'Indicators', 'HDGroups', 'HDRegions', 'Dimensions','Indices'}

\item{apikey}{str}
}
\value{
a dataframe
}
\description{
Get metadata from the human development report API
}
\examples{
country_list = get_meta(value = 'Countries', apikey = HDR_API)


}
